% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_one.R
\name{add_one}
\alias{add_one}
\title{A function which adds 1 to a number}
\usage{
add_one(x)
}
\arguments{
\item{x}{The number you would like to add one to.}
}
\description{
This function allows you to add 2 to a number.
}
\examples{
add_one(x = 14.1)
}
\keyword{adding}
